Chat-with-PDFs/requirements.txt
LangChain===0.0.184
PyPDF2
python-dotenv
streamlit==1.18.1
openai
faiss-cpu
altair
tiktoken
langchain-community
langchain_openai
langchain_text_splitters
langchain-community
urllib3==1.25.11
# uncomment to use huggingface llms
huggingface-hub

# uncomment to use instructor embeddings
InstructorEmbedding
sentence-transformers